db: Migrate last_read_at to float for sub-seconds
This fixes a bug where two sync pushes within one second cause a UNIQUE
key constaint error between documents and document_history.

I though I had fixed this by using a transaction, but it was in fact a
timing bug.

+++ internal/legacy/database_models.go
@@ -20,7 +20,7 @@ type DocumentData struct {
 
 type FileData struct {
 	ProgressData
-	DocumentId string `json:"document"`
-	Timestamp  int64  `json:"timestamp"`
-	PrettyName string `json:"pretty_name"` // User given name of Document, set via WebUI
+	DocumentId string  `json:"document"`
+	Timestamp  float64 `json:"timestamp"`
+	PrettyName string  `json:"pretty_name"` // User given name of Document, set via WebUI
 }
